fastAPI-graphql-mysql/requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cors==1.0.1
cross-web==0.4.1
cryptography==46.0.5
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.135.1
filelock==3.25.2
future==1.0.0
gevent==25.9.1
graphql-core==3.2.8
greenlet==3.3.2
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
passlib==1.7.4
pwdlib==0.3.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
pyotp==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
qrcode==8.2
requests==2.32.5
requests-file==3.0.1
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
strawberry-graphql==0.311.1
tldextract==5.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
Werkzeug==3.1.6
zope.event==6.1
zope.interface==8.2
